renamed main.py to service.py and fixed import
