+++ backend/python_service/test/test_service.py
@@ -1,12 +1,11 @@
 import unittest
+from backend.python_service.service import main_func
 
 
 class MainTest(unittest.TestCase):
 
     # Returns True or False.
     def test_main_func(self):
-        from ...python_service.main import main_func
-
         self.assertRaises(Exception, main_func())
